refactor: domain로직 분리
