+++ src/Components/Form/useIsValid.ts
@@ -1,3 +1,4 @@
+import { isCVCValid, isIssuerValid, isNumberValid, isOwnerValid, isPeriodValid } from "../../domainUtils";
 import useContextWrapper from "../../hooks/useContextWrapper";
 import {
   CardCVCContext,
@@ -17,71 +18,7 @@ import {
   CardValidityPeriodErrorContext,
 } from "../../routes/Payments/FormContextProvider";
 
-export const isNumberValid = (cardNumbers: CardNumbers, numberError: CardNumbersError) => {
-  if (
-    cardNumbers.firstNumbers?.length === 4 &&
-    cardNumbers.secondNumbers?.length === 4 &&
-    cardNumbers.thirdNumbers?.length === 4 &&
-    cardNumbers.fourthNumbers?.length === 4 &&
-    !numberError.firstNumbers?.isError &&
-    !numberError.secondNumbers?.isError &&
-    !numberError.thirdNumbers?.isError &&
-    !numberError.fourthNumbers?.isError
-  ) {
-    return true;
-  }
-  return false;
-};
-
-export const isIssuerValid = ({ name }: CardIssuer, { name: nameError }: CardIssuerError) => {
-  if (name && !nameError.isError) {
-    return true;
-  }
-  return false;
-};
-
-const PERIOD_ERROR_MESSAGE = {
-  NOT_VALID_YEAR: "잘못된 년도를 입력하셨습니다. 유효한 년도를 입력해주세요.",
-  NOT_VALID_MONTH: "잘못된 월을 입력하셨습니다. 유효한 월을 입력해주세요.",
-};
-
-export const isPeriodValid = (
-  { month, year }: CardValidityPeriod,
-  { month: monthError, year: yearError }: CardValidityPeriodError
-): { isValid: boolean; name?: "year" | "month"; errorMessage?: string } => {
-  const currentDate = new Date(Date.now());
-  const currentMonth = currentDate.getMonth() + 1;
-  const currentYear = Number(currentDate.getFullYear().toString().slice(2));
-
-  if (currentYear > Number(year)) {
-    return { isValid: false, name: "year", errorMessage: PERIOD_ERROR_MESSAGE.NOT_VALID_YEAR };
-  }
-
-  if (month?.length === 2 && year?.length === 2 && !monthError?.isError && !yearError?.isError) {
-    if (currentYear === Number(year) && currentMonth > Number(month)) {
-      return { isValid: false, name: "month", errorMessage: PERIOD_ERROR_MESSAGE.NOT_VALID_MONTH };
-    }
-    return { isValid: true };
-  }
-
-  return { isValid: false };
-};
-
-export const isOwnerValid = ({ name }: CardOwnerInfo, { name: errorName }: CardOwnerInfoError) => {
-  if (name && !errorName?.isError) {
-    return true;
-  }
-  return false;
-};
-
-export const isCVCValid = ({ value }: CardCVC, { value: valueError }: CardCVCError) => {
-  if (value?.length === 3 && !valueError.isError) {
-    return true;
-  }
-  return false;
-};
-
-export const isPasswordValid = ({ value }: CardPassword, { value: valueError }: CardPasswordError) => {
+const isPasswordValid = ({ value }: CardPassword, { value: valueError }: CardPasswordError) => {
   if (value?.length === 2 && !valueError.isError) {
     return true;
   }
